+++ CHANGELOG.md
@@ -8,3 +8,4 @@ HEAD
 * Add platformio library.json
 * Add first MQTTClient class and include ArduinoJson lib dep
 * Add release github action and build scripts
+* Support for esp8266 target using PubSubClient, WebSocketStreamClient and ArduinoHttpClient for WebSockets client